Code/requirements.txt
googletrans==3.1.0a0
pytorch-pretrained-bert
transformers
pytorch
numpy
nltk
gensim
fasttext
gdown